Support creating the wheel by building from source or repackaging binaries.
A top-level CMakeLists.txt has been added, it provides the
option BUILD_CMAKE_FROM_SOURCE defaulting to ON on Linux and OFF on
MacOSX and Windows.

If ON: CMake is build from source

If OFF: CMake binaries matching the host platform are downloaded and
packaged into a wheel.

Default can be overridden using a command like this one:

  python setup.pt bdist_wheel -- -DBUILD_CMAKE_FROM_SOURCE:BOOL=ON

See comments added to CMakeLists.txt for more details.