Replace usage of internal pip function `parse_requirements`
Starting with pip 10 the `parse_requirement` function was explicitly
moved to an internal namespace. In other words, using it is fraught with
problems since there are no guarantees it won't move again, have its
signature altered, or otherwise. Because of this, replace the use of the
internal function with an equivalent function.